+++ src/mappings/prices.ts
@@ -2,7 +2,7 @@ import { EventHandlerContext } from '../processor'
 import { network } from '../config'
 import { getOrCreateOraclePrice } from '../entities/oraclePrice'
 import { hexToString } from '@polkadot/util'
-import { decodeEvent } from '../eventsAndStorageSelector'
+import { decodeEvent } from '../types/eventsAndStorageSelector'
 export async function handleUpdatedPrices(ctx: EventHandlerContext) {
     let event = await decodeEvent(
         network,
