+++ src/utils/token.ts
@@ -4,7 +4,7 @@ import { codec } from '@subsquid/ss58'
 import { config, network } from '../config'
 import { invert } from 'lodash'
 import { hexToU8a } from '@polkadot/util'
-import { getVersionedStorage } from '../eventsAndStorageSelector'
+import { getVersionedStorage } from '../types/eventsAndStorageSelector'
 export const currencyKeyMap: { [index: number]: string } = {
     0: 'Native',
     1: 'XCM',
